requirements.txt
absl-py==2.1.0
agent==0.1.3
aiohappyeyeballs==2.4.3
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.6.2.post1
arxiv==2.1.3
arxiv2text==0.1.14
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitarray==3.0.0
bleach==6.1.0
blinker==1.8.2
boto3==1.35.1
botocore==1.35.99
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.31.4
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deepface==0.0.93
distlib==0.3.9
distro==1.9.0
dlib @ file:///C:/Users/Asus/Desktop/Project/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
docstring_parser==0.16
earthengine-api==0.1.416
emoji==2.12.1
entrypoints==0.4
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==27.0.0
fastapi==0.115.8
favicon==0.7.0
feedparser==6.0.11
filelock==3.15.4
fire==0.7.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-speech==2.30.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.8.4
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
helper==2.5.0
hexbytes==1.3.0
htbuilder==0.6.2
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
identify==2.6.6
idna==3.7
imageio==2.35.0
imbalanced-learn==0.12.3
imblearn==0.0
iniconfig==2.0.0
instructor==1.7.2
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.17
keras==3.4.1
kiwisolver==1.4.5
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-ollama==0.2.3
langchain-text-splitters==0.3.5
langsmith==0.3.4
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.43.0
lxml==5.3.0
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.26.0
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.4.0
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
openai==1.61.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.15
ospa==0.0.1
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.4
pdfminer.six==20240706
pgvector==0.3.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
pre_commit==4.1.0
preprocessor==1.1.3
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cord==2.6.1
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyheimdall==1.0.1
pyheimdall-heurist==0.1.0
pymdown-extensions==10.9
PyMuPDF==1.25.2
pyparsing==3.1.2
PyPDF2==3.0.1
pypiwin32==223
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==21.10
pyttsx3==2.98
pytz==2024.1
pyunormalize==16.0.0
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
ratelim==0.1.6
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==13.7.1
rlp==4.0.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.4
scikeras==0.13.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==75.8.0
sgmllib3k==1.0.0
shap==0.46.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.37
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.45.3
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
sympy==1.13.0
tbb==2021.13.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.8.10
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
tweepy==4.14.0
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
uritools==4.0.3
urlextract==1.9.0
urllib3==2.2.2
uvicorn==0.34.0
validators==0.33.0
virtualenv==20.29.1
watchdog==4.0.1
wcwidth==0.2.13
web3==7.6.1
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.3
wheel==0.45.1
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.17.2
zstandard==0.23.0
